--- app/src/main/java/com/example/alarmconnect/ui/circulo/CirculoFragment_bkp.kt
@@ -1,42 +1,0 @@
-package com.example.alarmconnect.ui.circulo
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
-import com.example.alarmconnect.databinding.FragmentCirculoBinding
-
-class CirculoFragment_bkp : Fragment() {
-
-    private var _binding: FragmentCirculoBinding? = null
-
-    // This property is only valid between onCreateView and
-    // onDestroyView.
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        val circuloViewModel =
-            ViewModelProvider(this).get(CirculoViewModel::class.java)
-
-        _binding = FragmentCirculoBinding.inflate(inflater, container, false)
-        val root: View = binding.root
-
-        //val textView: TextView = binding.textNotifications
-        //circuloViewModel.text.observe(viewLifecycleOwner) {
-        //    textView.text = it
-        //}
-        return root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
