+++ app/src/main/res/drawable/arrow_back.xml
@@ -1,0 +1,5 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="30dp" android:viewportHeight="512" android:viewportWidth="448" android:width="25dp">
+      
+    <path android:fillColor="#4E4E4E" android:pathData="M9.4,233.4c-12.5,12.5 -12.5,32.8 0,45.3l160,160c12.5,12.5 32.8,12.5 45.3,0s12.5,-32.8 0,-45.3L109.2,288 416,288c17.7,0 32,-14.3 32,-32s-14.3,-32 -32,-32l-306.7,0L214.6,118.6c12.5,-12.5 12.5,-32.8 0,-45.3s-32.8,-12.5 -45.3,0l-160,160z"/>
+    
+</vector>
