+++ app/src/main/res/layout/fragment_seleccionar_contacto.xml
@@ -5,19 +5,24 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.circulo.CirculoFragment">
+    android:gravity="center_horizontal"
+    android:orientation="vertical"
+    tools:context=".ui.circulo.SeleccionarContactoFragment">
 
-    <!-- Botón para volver -->
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorSurface"
-        app:navigationIcon="@drawable/ic_play"
-        app:title="Seleccionar Contacto de Confianza"
+    <TextView
+        android:id="@+id/titleSeleccionarContacto"
+        style="@style/TextAppearance.Material3.HeadlineMedium"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
+        android:text="Seleccionar Contacto"
+        android:textAlignment="center"
+        android:gravity="center"
+
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"/>
+
 
     <!-- RecyclerView para la lista de contactos -->
     <androidx.recyclerview.widget.RecyclerView
@@ -25,7 +30,7 @@
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_margin="16dp"
-        app:layout_constraintTop_toBottomOf="@id/toolbar"
+        app:layout_constraintTop_toBottomOf="@id/titleSeleccionarContacto"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintBottom_toTopOf="@id/buscarExternosButton"/>
