+++ app/src/main/res/layout/list_item_contacto.xml
@@ -11,7 +11,7 @@
         android:id="@+id/iconoUsuario"
         android:layout_width="40dp"
         android:layout_height="40dp"
-        android:src="@drawable/ic_play"
+        android:src="@drawable/baseline_account_circle_24"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"/>
@@ -22,7 +22,7 @@
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:text="Felipe Gomez"
-        android:textSize="18sp"
+        android:textSize="16sp"
         app:layout_constraintStart_toEndOf="@id/iconoUsuario"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintEnd_toStartOf="@id/seleccionarButton"/>
@@ -33,7 +33,7 @@
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:text="Hermano"
-        android:textSize="14sp"
+        android:textSize="13sp"
         android:textColor="@android:color/darker_gray"
         app:layout_constraintStart_toEndOf="@id/iconoUsuario"
         app:layout_constraintTop_toBottomOf="@id/nombreContacto"
@@ -46,7 +46,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Seleccionar"
-        android:textSize="14sp"
+        android:textSize="11sp"
         android:textColor="@android:color/white"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
